Bake skinned pose-union bounds into imported scene geometry
Extracts the offline bounds analysis (static primitive AABBs and the
skinned pose-union bake, which samples every animation driving a skin's
joints) into a shared module both emitters use. Imported scene geometry
referenced by a skinned node now carries the pose-union AABB instead of
its rest AABB, which under-covered the mesh once joints animate; a
primitive whose pose union cannot be computed carries no bounds and
renders unculled. A mesh shared by skinned and unskinned nodes carries
the union of both extents. Bumps the build-cache revision so cached
conversions rebuild with the new bounds.
